+++ src/main/java/insane96mcp/enhancedai/modules/skeleton/shoot/EARangedBowAttackGoal.java
@@ -107,7 +107,9 @@ public class EARangedBowAttackGoal extends EARangedAttackGoal<AbstractSkeleton> 
 		double dirZ = target.getZ() - entity.getZ();
 		double distanceXZ = Math.sqrt(dirX * dirX + dirZ * dirZ);
 		double yPos = target.getY(0d);
-		yPos += target.getEyeHeight() * 0.5 + (distanceY / distanceXZ);
+		yPos += target.getEyeHeight() * 0.5;
+		if (distanceXZ != 0f)
+			yPos += (distanceY / distanceXZ);
 		double dirY = yPos - abstractarrowentity.getY();
 		abstractarrowentity.shoot(dirX, dirY + distanceXZ * 0.17d, dirZ, f * 1.1f + ((float)distance / 32f) + (float)Math.max(distanceY / 48d, 0f), this.inaccuracy);
 		entity.playSound(SoundEvents.SKELETON_SHOOT, 1.0F, 1.0F / (entity.getRandom().nextFloat() * 0.4F + 0.8F));